src/requirements.txt
langchain-core==0.3.35
langchain==0.3.18
langchain-community==0.3.17
langchain-text-splitters==0.3.6
PyPDF2==3.0.1
databricks-vectorsearch==0.49
sentence-transformers==3.4.1
flashrank==0.2.8
databricks_langchain==0.3.0
mlflow==2.20.2